fixed cmp.go

+++ collector/info_schema_innodb_cmp.go
@@ -56,8 +56,8 @@ func ScrapeInfoSchemaInnodbCompression(db *sql.DB, ch chan<- prometheus.Metric) 
 			return err
 		}
 		ch <- prometheus.MustNewConstMetric(
-			infoSchemaInnodbCompressionInfoDesc, prometheus.GaugeValue,
-			uint32(page_size), uint32(compress_time), uint32(compress_ops_ok), uint32(uncompress_time), uint32(uncompress_ops), uint32uint32,
+			infoSchemaInnodbCompressionInfoDesc, prometheus.GaugeValue,float64(page_size),
+			compress_time, compress_ops_ok, uncompress_time, uncompress_ops, uint32uint32,
 		)
 	}
